add index.php e pequenas mudançãs em style.css

+++ _bd/cancelar_incrição.php
@@ -1,17 +1,22 @@
 <?php
+	session_start();
+	if (isset($_SESSION['usuario'])){
+		$id_evento = $_GET['id'];
 
-	$id_evento = $_GET['id'];
+		include "conexao.php";
 
-	include "conexao.php";
+		$conn = conectar();
 
-	$conn = conectar();
+		$sql = "DELETE FROM inscricao WHERE cpf_atleta=".$_SESSION['cpf']." and id_evento=".$id_evento";";
 
-	$sql = "DELETE FROM inscricao WHERE cpf_atleta=".$_SESSION['cpf']." and id_evento=".$id_evento";";
+		if ($conn->query($sql)){
+			header('Location: ../index.php');
+		}
 
-	if ($conn->query($sql)){
-		header('Location: incricao.php');
+		desconectar($conn);
+	} else {
+		header('Location: ../index.php')
 	}
-
-	desconectar($conn);
+	
 
 ?>
